news-aggregator/requirements.txt
requests==2.31.0
alpaca-py==0.21.0
aiohttp==3.9.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
